refactor: split explorer component (#1887)
