+++ packages/frontend/src/components/Explorer/Explorer.styles.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const Divider = styled.div`
+    padding-top: 10px;
+`;
