+++ packages/frontend/src/components/Explorer/FiltersCard/FiltersCard.styles.ts
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const CardHeader = styled.div`
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+
+    h5 {
+        margin: 0;
+        padding: 0;
+    }
+`;
