+++ packages/frontend/src/components/Explorer/ResultsCard/ResultsCard.styles.ts
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+
+export const CardHeader = styled.div`
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+`;
+
+export const CardHeaderLeftContent = styled.div`
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+
+    h5 {
+        margin: 0;
+        padding: 0;
+        margin-right: 10px;
+    }
+`;
+export const CardHeaderRightContent = styled.div`
+    display: flex;
+    align-items: center;
+    margin-right: 10px;
+    gap: 10px;
+`;
