+++ packages/frontend/src/components/Explorer/VisualizationCard/VisualizationCard.styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const CardHeader = styled.div`
+    display: flex;
+    flexdirection: row;
+    alignitems: center;
+`;
