+++ packages/frontend/src/pages/Explorer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Explorer } from '../components/Explorer';
+import Explorer from '../components/Explorer';
 import ExploreSideBar from '../components/Explorer/ExploreSideBar/index';
 import {
     useExplorerRoute,
